make issue plugin create/link/unlink actions overridable
